+++ docs/changelog.md
@@ -6,13 +6,14 @@
 
 - [[#138]](https://github.com/ewels/MegaQC/issues/138) Added `USER_REGISTRATION_APPROVAL` as a config variable, which defaults to true. This means that the admin must explicitly activate new users in the user management page (`/users/admin/users`) before they can login. To disable this feature, you need to create a config file (for example `megaqc.conf.yaml`) with the contents:
   ```yaml
-  STRICT_REGISTRATION: true
+  STRICT_REGISTRATION: false
   ```
   Then, whenever you run MegaQC, you need to `export MEGAQC_CONFIG /path/to/megaqc.conf.yaml
 
 ### New Features
 
 - [[#140]](https://github.com/ewels/MegaQC/issues/140) Added a changelog. It's here! You're reading it!
+- Tests for the REST API permissions
 
 ### Bug Fixes